refactor(st-09070): modularize multi-agent utilities
